Add dracut.conf.d/99-nvidia.conf file
Add dracut.conf.d/99-nvidia.conf file enforcing that the nvidia modules never
get added to the initramfs.

This is necessary because the userspace bits and the kernel modules must
be 100% in sync, otherwise things might break. Normally we build the initrd
before building the dkms / akmod so this issue will only happen if the initrd
somehow gets re-generated afterwards, still this is something which we should
avoid.

Patch from Hans de Goede